Update README and create html and css files
In an attempt to test out CSS styling away from the recipe repo, I have
decided to create a random site here to play around. Included here is
the creation of the base index.html and styles.css files, along with an
update of the README to serve as a reminder of the goal.